:recycle: Add example VM
