+++ CalculatorApp/Model/CalculationOperator.swift
@@ -1,0 +1,17 @@
+//
+//  CalculationOperator.swift
+//  CalculatorApp
+//
+//  Created by Shun Usami on 2019/10/03.
+//  Copyright © 2019 RyoIida. All rights reserved.
+//
+
+import Foundation
+
+enum CalculatorOperator: String {
+    case multiply = "×"
+    case divide = "÷"
+    case add = "+"
+    case sub = "-"
+    case equal = "="
+}
