+++ CalculatorApp/Model/Calculator.swift
@@ -1,0 +1,16 @@
+//
+//  Calculator.swift
+//  CalculatorApp
+//
+//  Created by Shun Usami on 2019/10/03.
+//  Copyright © 2019 RyoIida. All rights reserved.
+//
+
+import Foundation
+
+struct Calculator {
+    static func calculate(numbers: [String], operators: [CalculatorOperator]) -> String {
+        // TODO: Implement this!
+        return ""
+    }
+}
